refactor: enviando ButtonProps para arquivo externo
