+++ frontend/src/components/Button/Button.tsx
@@ -1,22 +1,6 @@
 import React, { ButtonHTMLAttributes, ReactNode } from 'react';
+import { ButtonProps } from './Button.types';
 import Spinner  from '../Spinner/Spinner';
-
-type ButtonVariant = 'primary' | 'secondary' | 'outline' | 'danger' | 'success';
-type ButtonSize = 'sm' | 'md' | 'lg';
-
-interface ButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
-  label?: string;
-  onClick?: () => void;
-  variant?: ButtonVariant;
-  disabled?: boolean;
-  loading?: boolean;
-  loadingText?: string;
-  icon?: ReactNode;
-  size?: ButtonSize;
-  className?: string;
-  style?: React.CSSProperties;
-  fullWidth?: boolean;
-}
 
 const Button: React.FC<ButtonProps> = ({
   label = '',
