+++ frontend/src/components/Button/Button.types.tsx
@@ -1,0 +1,18 @@
+import React, { ButtonHTMLAttributes, ReactNode } from 'react';
+
+type ButtonVariant = 'primary' | 'secondary' | 'outline' | 'danger' | 'success';
+type ButtonSize = 'sm' | 'md' | 'lg';
+
+export interface ButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
+  label?: string;
+  onClick?: () => void;
+  variant?: ButtonVariant;
+  disabled?: boolean;
+  loading?: boolean;
+  loadingText?: string;
+  icon?: ReactNode;
+  size?: ButtonSize;
+  className?: string;
+  style?: React.CSSProperties;
+  fullWidth?: boolean;
+}
